Add getHydrationFormData and postNewHydration functions
